move refactor Manifest to its own file
