+++ bindings/netstandard/ElectionGuard/ElectionGuard.Shared/Manifest.cs
@@ -1,0 +1,154 @@
+﻿using System;
+using System.Runtime.InteropServices;
+
+namespace ElectionGuard
+{
+    using NativeManifest = NativeInterface.Manifest.ManifestHandle;
+    using NativeInternalManifest = NativeInterface.InternalManifest.InternalManifestHandle;
+    using NativeElementModP = NativeInterface.ElementModP.ElementModPHandle;
+    using NativeElementModQ = NativeInterface.ElementModQ.ElementModQHandle;
+
+    /// <summary>
+    /// Use this entity for defining the structure of the election and associated
+    /// information such as candidates, contests, and vote counts.  This class is
+    /// based on the NIST Election Common Standard Data Specification.  Some deviations
+    /// from the standard exist.
+    ///
+    /// This structure is considered an immutable input object and should not be changed
+    /// through the course of an election, as it's hash representation is the basis for all
+    /// other hash representations within an ElectionGuard election context.
+    ///
+    /// See: https://developers.google.com/elections-data/reference/election
+    /// </summary>
+    public class Manifest : DisposableBase
+    {
+        public unsafe string ElectionScopeId
+        {
+            get
+            {
+                var status = NativeInterface.Manifest.GetElectionScopeId(
+                    Handle, out IntPtr value);
+                if (status != Status.ELECTIONGUARD_STATUS_SUCCESS)
+                {
+                    Console.WriteLine($"Manifest Error ObjectId: {status}");
+                    return null;
+                }
+                return Marshal.PtrToStringAnsi(value);
+            }
+        }
+
+        internal unsafe NativeManifest Handle;
+
+        public unsafe Manifest(string json)
+        {
+            var status = NativeInterface.Manifest.FromJson(json, out Handle);
+            if (status != Status.ELECTIONGUARD_STATUS_SUCCESS)
+            {
+                Console.WriteLine($"Manifest Error Status: {status}");
+            }
+        }
+
+        protected override unsafe void DisposeUnmanaged()
+        {
+            base.DisposeUnmanaged();
+
+            if (Handle == null || Handle.IsInvalid) return;
+            Handle.Dispose();
+            Handle = null;
+        }
+
+        public unsafe ElementModQ CryptoHash()
+        {
+            var status = NativeInterface.Manifest.CryptoHash(
+                    Handle, out NativeElementModQ value);
+            if (status != Status.ELECTIONGUARD_STATUS_SUCCESS)
+            {
+                Console.WriteLine($"CryptoHash Error Status: {status}");
+                return null;
+            }
+            return new ElementModQ(value);
+        }
+
+        public unsafe string ToJson()
+        {
+            var status = NativeInterface.Manifest.ToJson(
+                Handle, out IntPtr pointer, out UIntPtr size);
+            if (status != Status.ELECTIONGUARD_STATUS_SUCCESS)
+            {
+                Console.WriteLine($"ToJson Error Status: {status}");
+                return null;
+            }
+            var json = Marshal.PtrToStringAnsi(pointer);
+            return json;
+        }
+    }
+
+    /// <summary>
+    /// `InternalManifest` is a subset of the `Manifest` structure that specifies
+    /// the components that ElectionGuard uses for conducting an election.  The key component is the
+    /// `contests` collection, which applies placeholder selections to the `Manifest` contests
+    /// </summary>
+    public class InternalManifest : DisposableBase
+    {
+        /// <summary>
+        /// The hash of the election manifest
+        /// </summary>
+        public unsafe ElementModQ ManifestHash
+        {
+            get
+            {
+                var status = NativeInterface.InternalManifest.GetManifestHash(
+                    Handle, out NativeElementModQ value);
+                if (status != Status.ELECTIONGUARD_STATUS_SUCCESS)
+                {
+                    Console.WriteLine($"ManifestHash Error Status: {status}");
+                    return null;
+                }
+                return new ElementModQ(value);
+            }
+        }
+
+        internal unsafe NativeInternalManifest Handle;
+
+        public unsafe InternalManifest(Manifest election)
+        {
+            var status = NativeInterface.InternalManifest.New(election.Handle, out Handle);
+            if (status != Status.ELECTIONGUARD_STATUS_SUCCESS)
+            {
+                Console.WriteLine($"InternalManifest Error Status: {status}");
+            }
+        }
+
+        public unsafe InternalManifest(string json)
+        {
+            var status = NativeInterface.InternalManifest.FromJson(json, out Handle);
+            if (status != Status.ELECTIONGUARD_STATUS_SUCCESS)
+            {
+                Console.WriteLine($"InternalManifest Error Status: {status}");
+            }
+        }
+
+        protected override unsafe void DisposeUnmanaged()
+        {
+            base.DisposeUnmanaged();
+
+            if (Handle == null || Handle.IsInvalid) return;
+            Handle.Dispose();
+            Handle = null;
+        }
+
+        public unsafe string ToJson()
+        {
+            var status = NativeInterface.InternalManifest.ToJson(
+                Handle, out IntPtr pointer, out UIntPtr size);
+            if (status != Status.ELECTIONGUARD_STATUS_SUCCESS)
+            {
+                Console.WriteLine($"ToJson Error Status: {status}");
+                return null;
+            }
+            var json = Marshal.PtrToStringAnsi(pointer);
+            return json;
+        }
+    }
+
+}
