+++ bindings/netstandard/ElectionGuard/Tests/ElectionGuard.Tests/ElectionGuard.Tests.projitems
@@ -13,5 +13,6 @@
     <Compile Include="$(MSBuildThisFileDirectory)TestElection.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)TestElGamal.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)TestEncrypt.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)TestManifest.cs" />
   </ItemGroup>
 </Project>
